requirements.txt
django
djangorestframework
Pillow
psycopg2-binary
python-decouple
django-filter
drf-yasg
djangorestframework-simplejwt
django-rest-passwordreset
django-cors-headers
celery
redis
coreapi
docker
docker-compose
gunicorn
phonenumbers


